Move 'set app offline state' back to WebHost level
