+++ src/WebJobs.Script.WebHost/Controllers/HostController.cs
@@ -15,7 +15,6 @@ using Microsoft.AspNetCore.Mvc.WebApiCompatShim;
 using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.Azure.WebJobs.Host;
 using Microsoft.Azure.WebJobs.Host.Executors;
-using Microsoft.Azure.WebJobs.Script.Eventing.File;
 using Microsoft.Azure.WebJobs.Script.WebHost.Authentication;
 using Microsoft.Azure.WebJobs.Script.WebHost.Filters;
 using Microsoft.Azure.WebJobs.Script.WebHost.Management;
@@ -212,7 +211,7 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Controllers
                 }
 
                 // we're currently offline and the request is to bring the host back online
-                await FileChangeHelper.SetAppOfflineState(_applicationHostOptions.Value.ScriptPath, false);
+                await FileMonitoringService.SetAppOfflineState(_applicationHostOptions.Value.ScriptPath, false);
             }
             else if (desiredState == ScriptHostState.Offline && currentState != ScriptHostState.Offline)
             {
@@ -222,7 +221,7 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Controllers
                 }
 
                 // we're currently online and the request is to take the host offline
-                await FileChangeHelper.SetAppOfflineState(_applicationHostOptions.Value.ScriptPath, true);
+                await FileMonitoringService.SetAppOfflineState(_applicationHostOptions.Value.ScriptPath, true);
             }
             else
             {
