+++ src/WebJobs.Script/Eventing/File/FileChangeHelper.cs
@@ -11,30 +11,6 @@ namespace Microsoft.Azure.WebJobs.Script.Eventing.File
 {
     internal static class FileChangeHelper
     {
-        internal static async Task SetAppOfflineState(string rootPath, bool offline)
-        {
-            string path = Path.Combine(rootPath, ScriptConstants.AppOfflineFileName);
-            bool offlineFileExists = System.IO.File.Exists(path);
-
-            if (offline && !offlineFileExists)
-            {
-                // create the app_offline.htm file in the root script directory
-                string content = FileUtility.ReadResourceString($"{ScriptConstants.ResourcePath}.{ScriptConstants.AppOfflineFileName}");
-                await FileUtility.WriteAsync(path, content);
-            }
-            else if (!offline && offlineFileExists)
-            {
-                // delete the app_offline.htm file
-                await Utility.InvokeWithRetriesAsync(() =>
-                {
-                    if (System.IO.File.Exists(path))
-                    {
-                        System.IO.File.Delete(path);
-                    }
-                }, maxRetries: 3, retryInterval: TimeSpan.FromSeconds(1));
-            }
-        }
-
         internal static void TraceFileChangeRestart(ILogger logger, string changeDescription, string changeType, string path, bool isShutdown)
         {
             string fileChangeMsg = string.Format(CultureInfo.InvariantCulture, "{0} change of type '{1}' detected for '{2}'", changeDescription, changeType, path);
